Refactoring on the right way
Consolidated MenuItems and SubMenus. For now, menuconfig will be treated
as singular items again. Also search() now returns an iterator. This
makes searching logic much simpler, but requires that we re-create the
tree (folding together identical parent nodes on the way) after we've
found all matching leaves. So, therefore, this commit breaks some things
but will ultimately lead to a better final product in the end.